requirements.txt
discord.py
aiohttp
python-dotenv
flask
flask-cors
requests
